feat: complete delay queue && push lottery message

+++ server/idl/base.thrift
@@ -26,5 +26,5 @@ struct Gift{
     2:required i64 live_id (go.tag = "gorm:\"index\"")
     3:required string gift
     4:required i32 count
-    5:required i32 end_time
+    5:required i64 end_time
 }
